+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* DEPENDENCIES
+  * Updated ReverseDSC to version 2.0.0.16.
+
 # 1.23.621.1
 
 * AADAdministrativeUnit